完成新标签页的JSON、历史记录功能重构为Vue.js
TODO 书签部分

+++ src/js/new_tab.js
@@ -55,41 +55,37 @@ console.log('new_tab.js loaded')
         (() => {
             const bookmarks_dom = document.querySelector('#bookmarks ul')
 
-            let bookmarks
-
             // 获取所有收藏夹数据 https://developer.chrome.com/extensions/bookmarks
             chrome.bookmarks.getTree(root => {
                 // console.log(root)
 
-                const top_trees = root[0].children
-                // console.table('top_trees: ', top_trees)
-
-                // 轮流输出各顶级书签文件夹
-                top_trees.forEach(tree => parse_list(tree))
+                // 开始迭代处理根节点
+                root.shift().children.forEach((node, index) => parse_node(node, index))
             })
 
-            // 处理列表
-            const parse_list = list => {
-                // console.log('parse_list: ', list)
+            // 解析节点
+            const parse_node = (node, index) => {
+                // console.log('parse_node: ', node)
 
                 // 若为列表，添加标题
-                bookmarks_dom.innerHTML += '<li>&nbsp;[子收藏夹] ' + list.title + '( ' + list.children.length + ' 项)' + '</li>'
+                bookmarks_dom.innerHTML += '<li style="margin-top:2rem;">&nbsp;[子收藏夹] ' + node.title + '( ' + node.children.length + ' 项)' + '</li>'
 
-                list.children.forEach(item => {
-                    if (item.children) {
-                        // 若为列表，解析列表
-                        if (item.children.length > 0) parse_list(item.children)
+                // 迭代处理子节点
+                node.children.forEach(node => {
+                    if (node.children) {
+                        // 若为列表，继续解析节点
+                        if (node.children.length > 0) parse_node(node.children)
 
                     } else {
                         // 若为单项，解析单项
-                        parse_single(item)
+                        parse_single(node)
                     }
                 })
 
             }
 
             // 处理单项
-            const parse_single = item => bookmarks_dom.innerHTML += `<li><a title="${item.title}" href="${item.url}" data-id="${item.id}" target="_blank"><img src="chrome://favicon/${item.url}" /> ${item.title} <small>ID ${item.id}</small></a></li>`
+            const parse_single = item => bookmarks_dom.innerHTML += `<li><a title="${item.title} | ID ${item.id}" href="${item.url}" target="_blank"><img src="chrome://favicon/${item.url}" /> ${item.title}</a></li>`
 
         })() // end load bookmarks
 
@@ -97,8 +93,6 @@ console.log('new_tab.js loaded')
         ;
         (() => {
             const list_dom = document.querySelector('#history ul')
-
-            let list
 
             // 删除某一URL的所有历史记录，并刷新页面
             const delete_item = url => {
@@ -119,13 +113,15 @@ console.log('new_tab.js loaded')
                         const dom_item = document.createElement('li')
 
                         // 链接
-                        const dom_link = `<a title="${item.title}" href="${item.url}" data-id="${item.id}" target="_blank"><img src="chrome://favicon/${item.url}" /> ${item.title} <small>ID ${item.id}</small></a>`
+                        const dom_link = `<a title="${item.title} | ID ${item.id}" href="${item.url}" target="_blank"><img src="chrome://favicon/${item.url}" /> ${item.title}</a>`
                         dom_item.innerHTML = dom_link
 
                         // 删除按钮
                         const dom_delete = document.createElement('span')
                         dom_delete.innerHTML = '<i class="fal fa-trash"></i>'
-                        dom_delete.addEventListener('click', () => delete_item(item.url))
+                        dom_delete.addEventListener('click', () => {
+                            if (confirm(`删除“${item.title}”的访问记录？`)) delete_item(item.url)
+                        })
                         dom_item.append(dom_delete)
 
                         // 组装单项DOM，并输出到容器DOM顶部
@@ -141,11 +137,9 @@ console.log('new_tab.js loaded')
                 parse_items
             )
 
-            // 清空已获得的浏览记录
+            // 清空所有浏览记录
             document.querySelector('#clear_history').addEventListener('click', () => {
-                chrome.history.deleteAll(() => {
-                    location.reload()
-                })
+                if (confirm('清空所有浏览历史记录？')) chrome.history.deleteAll(() => location.reload())
             })
 
         })() // end load history
